added delete approval dialog box before org delete

+++ src/app/views/admin/admin-orgs-mamagement-item/admin-orgs-mamagement-item.component.ts
@@ -1,6 +1,8 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {ImageService} from '../../../core/image.service';
 import {OrgService} from '../../organization/org.service';
+import {DeleteApproveComponent} from "../../../shared/delete-approve/delete-approve.component";
+import {MatDialog} from "@angular/material";
 
 @Component({
   selector: 'sk-admin-orgs-mamagement-item',
@@ -11,7 +13,8 @@ export class AdminOrgsMamagementItemComponent implements OnInit {
   @Input() org;
   logoUrl = '';
   constructor(public imgService: ImageService,
-              public orgService: OrgService) { }
+              public orgService: OrgService,
+              private dialog: MatDialog) { }
 
   ngOnInit() {
     this.imgService.getOrgLogo$(this.org.orgId)
@@ -22,6 +25,21 @@ export class AdminOrgsMamagementItemComponent implements OnInit {
   }
 
   deleteClicked(org) {
-    this.orgService.deleteOrg(org.orgId);
+    const dialogRef = this.dialog.open(DeleteApproveComponent, {
+      data: {
+        'orgId': org.orgId,
+        'verifyPhrase': org.orgId
+      },
+      height: '400px',
+      width: '600px',
+    });
+
+    dialogRef.afterClosed()
+      .subscribe(res => {
+        if (res) {
+          this.orgService.deleteOrg(org.orgId);
+        }
+      });
+
   }
 }
